pyproject.toml
[project]
name = "nsee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ffsedd", email = "ffsedd@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "scikit-image>=0.26.0",
]

[project.scripts]
nsee = "nsee.main:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
    "setuptools>=82.0.1",
    "wheel>=0.46.3",
]
[tool.setuptools.packages.find]
where = ["src"]
[tool.setuptools]
package-dir = {"" = "src"}
[tool.pytest.ini_options]
pythonpath = ["src"]
